pyproject.toml
[tool.poetry]
name = "autoblocksai"
# The version is not maintained here. It's set in the release workflow via the version input.
version = "0.0.0"
description = "Python client for Autoblocks"
license = "MIT"
authors = ["Autoblocks Engineering <engineering@autoblocks.ai>"]
readme = "README.md"
homepage = "https://github.com/autoblocksai/python-sdk"
repository = "https://github.com/autoblocksai/python-sdk"
documentation = "https://docs.autoblocks.ai"
packages = [{include = "autoblocks"}]

[tool.poetry.dependencies]
python = "^3.9.0"

# NOTE: These dependencies should be as relaxed as possible.
# See CONTRIBUTING.md for more information.
httpx = ">=0.24.0"
click = ">=8.0.0"
pyyaml = ">=6.0.0"
orjson = ">=3.0.0"
tenacity = ">=8.0.0"
cuid2 = ">=2.0.0"
opentelemetry-exporter-otlp-proto-http = ">=1.0.0"
opentelemetry-api = ">=1.0.0"
opentelemetry-sdk = ">=1.0.0"

[tool.poetry.group.dev.dependencies]
pre-commit = "^4.0.0"
pytest = "8.3.5"
pytest-httpx = "0.31.2"
freezegun = "^1.2.2"
pydantic = "^2.5.3"
mypy = "^1.8.0"
types-pyyaml = "^6.0.12.12"
types-orjson = "^3.6.2"
types-click = "^7.1.8"
flask = "^3.0.2"
gunicorn = ">=23,<23.1"
syrupy = "^4.6.1"
openai = "1.68.2"
ragas = "0.2.14"
datasets = "^3.0.0"
nltk = "^3.9.1"
rapidfuzz = "^3.10.0"
llama-index = "^0.12.0"

[tool.poetry.scripts]
prompts = "autoblocks._impl.prompts.cli.main:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120

[tool.ruff]
lint.select = ["E", "F", "I001"]
line-length = 120

[tool.ruff.lint.isort]
force-single-line = true
known-first-party = ["autoblocks"]

[tool.mypy]
python_version = "3.11"
strict = true
explicit_package_bases = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false
